Improve Error Handling & Allow Blast Radius Viewer to be Accessed Even If JSON Data Loading Failed
If the JSON data could not be parsed, then error messages will notify the user instead of silently failing.

In addition, if there was an error loading JSON data, the viewer can still be accessed. (The graph will still be there but the JSON data and some of the mouse events + colorization will be absent.